matchobj: make nfield, etc 32-bit rather than 16-bit (HST images can overflow this!)
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@15407 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ blind/matchobj.h
@@ -60,14 +60,14 @@ struct match_struct {
     // (only counts a single ABCD permutation)
     int16_t quad_npeers;
 
-	int16_t nmatch;
-	int16_t ndistractor;
-	int16_t nconflict;
-	int16_t nfield;
-	int16_t nindex;
+	int nmatch;
+	int ndistractor;
+	int nconflict;
+	int nfield;
+	int nindex;
 
 	// nbest = besti+1 = nmatch + ndistractor + nconflict <= nfield
-	int16_t nbest;
+	int nbest;
 
 	float logodds;
